nfp: add app specific rx/tx metadata infrastructure
Some applications may want to have app specific metadata included on RX
or TX. A new meta data type is allocated at the end of the current meta
data type range for this purpose.

The app infrastructure provides this by exposing three new generic app
APIs, two of which are only applicable on RX, i.e. parsing and
populating the fields, and the other for prepending the meta data on TX.

The RX metadata structure available to apps between the parsing and
populating callbacks consist of a 8 bit descriptor and a further 64 bit
wide data field, both can be used and interpreted by the app. Though,
the following assumptions are made:

1) The infrastructure does assume that the descriptor is non-zero if
   there were any meta data parsed. Zero valued descriptors would be
   interpreted as no app meta data to set in the skb.

2) If 64 bits are not enough to store the app meta data - which is
   probably unlikely. It is large enough to store an additional buffer
   pointer. Any memory allocated here must be freed again when the
   skb_set_meta' callback is invoked. The app is wholely responsible for
   freeing any memory allocated.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>