nfp: mbl: UAL can decide whether VF reprs are needed
Some UAL applications may not have any need for VF representors, so
provide the ability for the UAL to select whether it wants the MBL to
spawn VF representors or not.

Note that SR-IOV VFs can only be created if the UAL is first registered,
so we will always have an answer for this to succeed.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>