nfp: mbl: provide means to set repr port ID
UAL modules can use the nfp_ual_set_port_id API to set the port ID for a
representor to its desired value.

The UAL only has access to set the lower 20bits of the port ID. Any
modification to the other 12bits will result in the operation being
rejected.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>