nfp: mbl: close all reprs when stopping the UAL
Since the UAL is in control of the firmware, when we stop the UAL, we
close all reprs to ensure the system stays in a sane state.

Generally, the UAL will only unload when the nfp module will also be
unloaded, but this doesn't have to be the case. Another UAL module may
be loaded afterward, so the reprs need to be closed.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>