nfp: mbl: provide SR-IOV functions
The SR-IOV enable/disable functions are passed through to the UAL
module, if it provides them. The MBL generates the VF reprs when SR-IOV
is enabled.

We don't allow the initialization of the MBL app when there are already
VFs configured. The UAL should always have the ability to react on
enabling/disabling SR-IOV VFs.

We could have gone the route of checking whether the UAL is already
available at this time, however, this is an odd use case so we just
reject this outright.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>