nfp: mbl: provide VLAN add/kill APIs
Pass through the RX VLAN ndo's to the UAL.

The UAL has the ability to set the port ID for a VLAN interface on a
specific repr. When VLAN traffic is being transmitted, the UAL provided
port ID is substituted for the repr port ID.

The VLAN metadata_dst entries are generated when the VLAN interface is
created and destroyed with the VLAN interface. The lower device for the
VLAN interface is always the same as the lower device for the repr at
the time of packet transmission.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>