nfp: mbl: allocate space for per repr private data
The MBL app is interested in per repr private data, so we allocate and
free this appropriately.

At the moment, we only make space for UAL priv data which it may choose
to use or not. The UAL is wholly responsible to allocate and free memory
it needs per repr during its init/clean callbacks.

We will extend the MBL to make use of the per repr priv data for its own
internal use later.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>