nfp: add app callbacks init_complete and clean_begin
The init_complete callback is executed after the app has started and
vNICs have been initialised. This is useful for apps which don't have
vNICs, but still may want to have an indication of when initialisation
is complete.

Similarly, during cleanup we want a callback before vNICs are being
cleaned so apps without vNICs may behave appropriately.

Note that subsequent patches will introduce the ability to have an
app without a vNIC.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>