nfp: probe devices that don't advertise vnics
We may have a device that forms part of a larger app environment, but
this device does not advertise any vnics.

To handle this case, we default to a max number of vnics of 0 instead of
1. If we have a device that has a maximum number of vnics of 0, then we
do a limited initialization which avoids all the vnic specific init
paths. Similarly on cleanup.

Changing the default is safe since the nfd_cfg_pfX_num_ports symbol ABI
has been around for a long time, and all upstream firmware images have
this defined.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>