nfp: add NFP_PORT_PHYS_PORT_EXP expander type
Although ports attached to an expander card, e.g. NICmod or port
expander, is exactly the same as a NFP_PORT_PHYS_PORT type, there are
subtle differences in behavior we want.

For example, stats won't be read from the BAR and we need a different
naming scheme for these ports.

The devlink functions, i.e. split/unsplit, has not been tested yet with
this change, so we could expect some future improvements in this regard.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>