nfp: mbl: handle port expander devices
For now, we detect port expander's by means of the AMDA0997. This will
be improved in the future, where the BSP will make other identification
available.

We force the probing order here, since the nicmod/port expander will
already have firmware loaded and could get probed before the main NFP.

Note that the ctrl vNIC is only required on the main NFP.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>