nfp: mbl: implement the UAL free callback
The free callback is invoked in two cases, when the UAL unregisters
itself or when the global MBL context is about to be freed.

This should be used by the UAL module to free its cookie, or any other
resources that it used.

This is different to the clean callback because it could happen that a
device is unbound, thus causing the UAL to be stopped (but not
unregistered). Then the device can be added back again, which will
reinitialize the UAL module.

Conversely, at the point when the free callback is invoked, the UAL
cannot be reinitialized again without being registered again.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>