nfp: mbl: verify new MTU within TX vNIC MTU
The MBL knows which TX vNIC has been selected by the UAL module, so it
can and should verify that the new MTU isn't higher than the TX vNIC
acting as the lower device for the repr.

However, the UAL handles the RX vNIC without interaction by the MBL, so
the UAL module needs to make that verification on its own.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>