samples: ualt: encode the port expander cluster info
With the port control message, we encode the port expander cluster and
cluster port values into the port field.

Port expander firmware uses this value to identify ports. The UALT
sample firmware uses this scheme directly.

Note that we skip ports 0.6 and 0.7 when sending control messages. These
ports are currently unused on the port expander mockup and cause
interference since the channel_base number is always set to 0.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>