nfp: link repr netdev features with vNICs
Some repr netdev features make sense to have linked to the vNIC, e.g.
checksum offloading and VLAN offloading.

The transport mechanism to indicate to the firmware to do this
offloading is implemented via the vNIC. So the repr can't have this
feature enabled if the vNIC doesn't support it.

Furthermore, if the vNIC features are modified, the app needs to take an
action to synchronize these changes to its representors. This can
argueably be a core feature of the driver, but there are cases where the
app needs to make certain app specific decisions on this event.

There is still a piece missing here. We should prevent changes made to
the representor features from circumventing the synchronization between
the vNIC and repr features enforced by the app.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>