nfp: mbl: sync vNIC features with reprs
Enforce a strict relationship between repr features and vNICs. Whenever
any available offload feature (only some are implemented at the moment)
becomes disabled on a vNIC, disable the feature on all reprs. And only
enable the feature on reprs once all vNICs have that feature available
again.

Reprs can easily be allocated to the different vNICs, so we need to
enforce this relationship.

We still need to ensure that users don't disable feature on reprs which
are still enabled on the vNIC. This won't have the desired effect.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>