nfp: implement the IPsec offload skeleton
IPsec offload, when available, is provided for both representors and
vNICs. Add the infrastructure to initialize and cleanup IPsec offload
references on these devices.

Store the IPsec data handle in the netdev priv structure for each type
of netdev.

Note that we disable IPsec offload support if either its not supported
by the kernel, or on kernels older than v4.14. The latter restriction is
added since a there is a bug in older kernels which cause negative ref
counts on netdevs when used for IPsec offload.

Signed-off-by: David George <david.george@netronome.com>
Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>