nfp: initialize IPsec only if supported by firmware
Use the vNIC extended capabilities to determine whether IPsec offload is
available by the firmware.

Only initialize the IPsec component on a particular netdev if it's
supported by firmware. If it is, then enable the ESP hardware offload
feature.

We support this on both representors and vNICs.

Note that for representors to support IPsec offload, the underlying vNIC
also needs to support this.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>