nfp: IPsec RX and TX paths
IPsec TX and RX paths make use of a NFD metadata slot, shared for two
exclusive uses for the respective direction.

On RX, the metadata contains the SA index value. If the SA index was
received and is validated, the packet has been successfully decrypted by
firmware.

On TX, the metadata contains the offload handle and sequence number,
forming an offload request.

Signed-off-by: David George <david.george@netronome.com>
Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>