nfp: provide ability to accumulate MAC stats
The port expander card does not have the capacity to run any MAC
statistic accumulation in firmware. Since this is a relatively low rate
platform, the driver can accumulate the statistics directly from the MAC
registers.

Provide the infrastructure to achieve this. There are no consumers of
this function yet.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>