samples: ual: clear MAC stats via debugfs
For each physical port representor, we provide the ability to clear the
MAC stats via debugfs. This is a useful feature for ODM testing that
will be using this module.

NOTE: This is only a debugging feature and is expected to be used while
the system is idle between tests. The clearing options are not atomic so
undefined results may occur if used while traffic is flowing through the
system.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>