nfp: mbl: add repr_xmit callback for the ual
This provides the UAL the ability to get a callback for each packet that
gets transmitted via a representor. The UAL can take some action if it
desires at this point.

The main purpose of this callback is for the UAL to be able to direct
traffic to vNICs on a per-packet basis, e.g. flow affinity to a
particular vNIC.

The precedence is, UAL gets first option on setting the destination
vNIC, then any preconfigured VLAN destinations and finally the default
vNIC is selected for the representor.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>