nfp: use QCP offset/size per device type
The queue controller (QCP) is accessed based on a device specific 
offset. The NFP3800 device also supports more queues.

Furthermore, the NFP3800 VFs also access the QCP differently to how the 
NFP6000 VFs accesses it, though still indirectly. Fortunately, we can 
remove the offset all together for both VF types. This is safe for 
NFP6000 VFs since the offset was effectively a wrap around and only used 
for convenience to have it set the same as the NFP6000 PF.

Adjust the QCP offset and size accordingly per device type.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>