samples: mbl: use debugfs to toggle/read app metadata
App metadata is enabled by default, however, this is can affect
performance and is not convenient for testing purposes. So instead,
provide a debugfs interface to enable TX metadata, and also provide a
debugfs interface to read the last RX metadata received and read/modify
the TX metadata to send.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>