nfp: attempt fw load from flash first
Whenever the hwinfo key, app_fw_from_flash is set to 1, the NFP has been
configured to always use firmware from flash. In that case, we try
loading stored firmware first before attempting the normal firmware
loading procedure.

This means that if firmware is available on flash, or already loaded
from flash, the driver will not attempt a soft-reset regardless of
whether another suitable firmware file was found on disk.

This is useful for port expanders which share AMDA numbers with the
main NFP or multi-host systems which always load firmware from flash.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>