nfp: mbl: support non-zero nicmod device index
The device index is currently calculated with the intent that the PCIe
unit number differentiates different devices, however, for nicmods this
won't be true since each nicmod will be connected to the host via PCIe
unit #0.

Store the device ID in the device context structure. For the main NFP
device, this will be set to the PCIe unit number (the original
behavior), but for nicmods the NFP ID hwinfo key will be used instead.

This does not alter existing behavior since up to now, only a single
nicmod was supported.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>