nfp: mbl: define the port expander port offsets
The main NFP accesses the port expander ports through a DSA mechanism.
The port numbering on the main NFP is based on the NSP eth_index with
an offset per port expander device.

Define the offsets that UAL applications can use to refer to the port
expander ports. This offset is defined per port expander slot.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>